Fixed bug creating FNT section of Nitro.Roms
I forgot to seek to the position where the table was
calculated to be.

* Removed tail data from ARM. It didn't work with some games
where there is no padding after the ARM file. Furthermore,
I think it's rubbish data from a previous ARM file. It could be that the SDK
recreate a game just changing the ARM (code file) so in the case
the new ARM is smaller there will be old, unused data.